requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.9.2
pydantic-settings==2.5.2

python-dotenv==1.0.1

chromadb==0.5.5

transformers==4.44.2
accelerate==0.34.2
sentencepiece==0.2.0

docling==2.4.1

aiosqlite==0.20.0

pytest==8.3.2
pytest-asyncio==0.21.1
onnxruntime==1.16.3
httpx==0.27.2
requests==2.32.3
psutil==6.0.0


